+++ modules/dependencies/lib/dependencies.dart
@@ -5,6 +5,7 @@ export 'package:return_success_or_error/return_success_or_error.dart';
 export 'package:get/get.dart';
 export 'package:font_awesome_flutter/font_awesome_flutter.dart';
 export 'package:logger/logger.dart';
+export 'package:flutter_overlay_loader/flutter_overlay_loader.dart';
 
 ///Pacakages Internos
 export 'package:core/core.dart';