+++ apps/client_android_modelviewlabsorg_desafio_api/lib/src/app_widget.dart
@@ -12,12 +12,24 @@ final class AppWidget extends StatelessWidget {
       title: 'Desafio - Model View Labs',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: const Color.fromARGB(255, 201, 193, 214)),
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: const Color.fromARGB(255, 0, 6, 59),
+          brightness: Brightness.light,
+          primary: const Color.fromARGB(255, 0, 6, 59),
+          secondary: const Color.fromARGB(255, 159, 167, 241),
+        ),
+        elevatedButtonTheme: ElevatedButtonThemeData(
+          style: ElevatedButton.styleFrom(
+              backgroundColor: const Color.fromARGB(255, 0, 6, 59),
+              textStyle: const TextStyle(
+                color: Color.fromARGB(255, 159, 167, 241),
+              )),
+        ),
         useMaterial3: true,
       ),
       binds: [
         ...CoreBinding().dependencies(),
-        ],
+      ],
       getPages: AppModule().routes,
     );
   }