+++ modules/core/lib/src/ui/widgets/logo_padrao.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class LogoPadrao extends StatelessWidget {
+  const LogoPadrao({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Image.asset(
+          'assets/logo.png',
+          height: 200,
+        ),
+        Text(
+          'Consumo API',
+          style: Theme.of(context).textTheme.headlineMedium,
+        )
+      ],
+    );
+  }
+}