+++ modules/dependencies/lib/dependencies.dart
@@ -6,6 +6,7 @@ export 'package:get/get.dart';
 export 'package:font_awesome_flutter/font_awesome_flutter.dart';
 export 'package:logger/logger.dart';
 export 'package:flutter_overlay_loader/flutter_overlay_loader.dart';
+export 'package:validatorless/validatorless.dart';
 
 ///Pacakages Internos
 export 'package:core/core.dart';