+++ modules/home/lib/src/ui/confirma_valida_senha/confirma_valida_senha_page.dart
@@ -1,5 +1,6 @@
 import 'package:dependencies/dependencies.dart';
 import 'package:flutter/material.dart';
+
 import 'confirma_valida_senha_controller.dart';
 
 class ConfirmaValidaSenhaPage extends GetView<ConfirmaValidaSenhaController> {
@@ -7,74 +8,77 @@ class ConfirmaValidaSenhaPage extends GetView<ConfirmaValidaSenhaController> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.amber,
-      body: Column(
+    return PageBase(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
         children: [
-          Center(
-            child: SizedBox(
-              width: 80,
-              height: 80,
-              child: IconButton(
-                  onPressed: () {
-                    // controller.signInGoogleLogin(
-                    //   onSuccess: () {
-                    //     Get.snackbar("Sucesso", "Sucesso ao fazer login");
-                    //   },
-                    //   onFail: () {
-                    //     Get.snackbar("Erro", "Erro ao fazer login");
-                    //   },
-                    // );
-                  },
-                  icon: const FaIcon(FontAwesomeIcons.google)),
+          const SizedBox(
+            height: 10,
+          ),
+          const VerificadoPadrao(),
+          const SizedBox(
+            height: 10,
+          ),
+          const Text('A senha foi validada com sucesso pela API!'),
+          Padding(
+            padding: const EdgeInsets.symmetric(
+              horizontal: 40,
+              vertical: 20,
+            ),
+            child: Column(
+              children: [
+                const SizedBox(
+                  height: 10,
+                ),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    ElevatedButtonPadrao(
+                      onPressed: () => Get.offAllNamed(Routes.initial.caminho),
+                      label: 'Teste Nova Senha',
+                    ),
+                  ],
+                ),
+              ],
             ),
           ),
-          Center(
-            child: SizedBox(
-              width: 80,
-              height: 80,
-              child: IconButton(
-                  onPressed: () {
-                    // controller.logOut(
-                    //   onSuccess: () {
-                    //     Get.snackbar("Sucesso", "Sucesso ao fazer LogOut");
-                    //   },
-                    //   onFail: () {
-                    //     Get.snackbar("Erro", "Erro ao fazer LogOut");
-                    //   },
-                    // );
-                  },
-                  icon: const FaIcon(FontAwesomeIcons.arrowRightFromBracket)),
-            ),
+          const SizedBox(
+            height: 20,
           ),
-          Center(
-            child: SizedBox(
-              width: 80,
-              height: 80,
-              child: IconButton(
-                  onPressed: () {
-                    // mostrarDialogoConfirmacao(onSuccess: () {
-                    //   controller.apagarConta(
-                    //     confirmacao: true,
-                    //     onSuccess: () {
-                    //       Get.snackbar("Sucesso", "Sucesso ao apagar a conta");
-                    //       Get.toNamed("/login");
-                    //     },
-                    //     onFail: () {
-                    //       Get.snackbar("Erro", "Erro ao apagar a conta");
-                    //     },
-                    //   );
-                    //   Get.back();
-                    // });
-                  },
-                  icon: const FaIcon(FontAwesomeIcons.userXmark)),
-            ),
-          ),
-          Center(
-            child: SizedBox(
-              width: 80,
-              height: 80,
-              child: Obx(() => Text(controller.teste.value)),
+          Expanded(
+            child: Container(
+              decoration: BoxDecoration(
+                color:
+                    Theme.of(context).colorScheme.inversePrimary.withAlpha(80),
+                border: Border(
+                  top: BorderSide(
+                    width: 2,
+                    color: Theme.of(context).colorScheme.secondary,
+                  ),
+                ),
+              ),
+              child: Column(
+                children: [
+                  const SizedBox(
+                    height: 30,
+                  ),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Obx(() => Text("id - ${controller.id}")),
+                    ],
+                  ),
+                  const SizedBox(
+                    height: 10,
+                  ),
+                   Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Obx(() => Text("message -${controller.message}")),
+                    ],
+                  ),
+                ],
+              ),
             ),
           ),
         ],
@@ -82,19 +86,3 @@ class ConfirmaValidaSenhaPage extends GetView<ConfirmaValidaSenhaController> {
     );
   }
 }
-
-// void _mostrarDialogoConfirmacao({
-//   required VoidCallback onSuccess,
-// }) {
-//   Get.defaultDialog(
-//     title: "Confirmação",
-//     middleText: "Tem certeza que deseja apagar sua conta?",
-//     textConfirm: "Sim",
-//     textCancel: "Não",
-//     confirmTextColor: Colors.white,
-//     onConfirm: onSuccess,
-//     onCancel: () {
-//       Get.back(); // Fecha o diálogo
-//     },
-//   );
-// }