+++ modules/core/lib/src/models/result_validate.dart
@@ -1,0 +1,14 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+
+import 'package:dependencies/dependencies.dart';
+
+
+final class ResultValidate {
+  final ValidateSuccessModel? validateSuccess;
+  final ErrorResponseModel? errorResponse;
+  ResultValidate({
+    this.validateSuccess,
+    this.errorResponse,
+  });
+  
+}