+++ modules/core/lib/src/ui/utilites/icons_padrao.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/widgets.dart';
+
+class IconsPadrao {
+  IconsPadrao._();
+
+  static const _kFontFam = 'IconsPadrao';
+  static const String? _kFontPkg = null;
+
+  static const IconData eye = IconData(0xf06e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData eyeSlash = IconData(0xf070, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+}