+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.3
 Desenvolvimento do layout.
 1 - Configuração core controller para disponibilizar os metodos de consumo da api.
+2 - Configuração do mixin loader.
+3 - Configuração do mixin messages.
 
 ## 0.0.2
 Desenvolvimento do código fonte.