+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Desenvolvimento do layout.
 1 - Configuração core controller para disponibilizar os metodos de consumo da api.
 2 - Configuração do mixin loader.
 3 - Configuração do mixin messages.
+4 - Configuração do thema.
+5 - Configuração dos assets.
+6 - Criação da tela ValidaSenhaPage.
 
 ## 0.0.2
 Desenvolvimento do código fonte.